# local/frontend/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentAuditAction": {
        "enum": [
          "INCIDENT_CREATED",
          "EVIDENCE_LINKED",
          "EVIDENCE_UNLINKED",
          "INCIDENT_SPLIT",
          "STATUS_CHANGED",
          "OWNER_CHANGED",
          "SEVERITY_CHANGED",
          "DUE_STATE_CHANGED",
          "DUE_TIME_CHANGED",
          "RECURRENCE_RECORDED",
          "HANDOVER_ACKNOWLEDGED"
        ],
        "title": "IncidentAuditAction",
        "type": "string"
      },
      "IncidentAuditOutput": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/IncidentAuditAction"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "id",
          "action",
          "occurred_at",
          "payload"
        ],
        "title": "IncidentAuditOutput",
        "type": "object"
      },
      "IncidentAuditResponse": {
        "additionalProperties": false,
        "properties": {
          "audit_events": {
            "items": {
              "$ref": "#/components/schemas/IncidentAuditOutput"
            },
            "title": "Audit Events",
            "type": "array"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "audit_events"
        ],
        "title": "IncidentAuditResponse",
        "type": "object"
      },
      "IncidentCollectionResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "IncidentCollectionResponse",
        "type": "object"
      },
      "IncidentDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "audit_events": {
            "items": {
              "$ref": "#/components/schemas/IncidentAuditOutput"
            },
            "title": "Audit Events",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "due_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due State"
          },
          "due_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Time"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "machine_id": {
            "format": "uuid",
            "title": "Machine Id",
            "type": "string"
          },
          "owner_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Ref"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "machine_id",
          "status",
          "owner_ref",
          "severity",
          "due_state",
          "due_time",
          "created_at",
          "updated_at",
          "audit_events"
        ],
        "title": "IncidentDetailResponse",
        "type": "object"
      },
      "IncidentEvidenceItem": {
        "additionalProperties": false,
        "properties": {
          "canonical_event_type": {
            "title": "Canonical Event Type",
            "type": "string"
          },
          "canonical_payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Canonical Payload",
            "type": "object"
          },
          "component_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component Id"
          },
          "deterministic_rule_identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deterministic Rule Identifier"
          },
          "evidence_id": {
            "format": "uuid",
            "title": "Evidence Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "link_id": {
            "format": "uuid",
            "title": "Link Id",
            "type": "string"
          },
          "link_reason": {
            "title": "Link Reason",
            "type": "string"
          },
          "linked_at": {
            "format": "date-time",
            "title": "Linked At",
            "type": "string"
          },
          "machine_id": {
            "format": "uuid",
            "title": "Machine Id",
            "type": "string"
          },
          "original_source_record_id": {
            "title": "Original Source Record Id",
            "type": "string"
          },
          "original_timestamp": {
            "format": "date-time",
            "title": "Original Timestamp",
            "type": "string"
          },
          "provenance": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Provenance",
            "type": "object"
          },
          "raw_source_payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Raw Source Payload",
            "type": "object"
          },
          "relationship_type": {
            "$ref": "#/components/schemas/IncidentEvidenceRelationshipType"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "unlinked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unlinked At"
          }
        },
        "required": [
          "link_id",
          "evidence_id",
          "is_active",
          "relationship_type",
          "deterministic_rule_identifier",
          "link_reason",
          "linked_at",
          "unlinked_at",
          "machine_id",
          "component_id",
          "source_type",
          "original_source_record_id",
          "original_timestamp",
          "canonical_event_type",
          "canonical_payload",
          "raw_source_payload",
          "provenance"
        ],
        "title": "IncidentEvidenceItem",
        "type": "object"
      },
      "IncidentEvidenceRelationshipType": {
        "description": "Controlled MVP vocabulary for why evidence is attached to an incident.\n\nRELATED is intentionally broad because the specification does not yet define\nfiner evidence roles. RECURRENCE and VERIFICATION are kept explicit because\nthey have lifecycle meaning in the MVP requirements.",
        "enum": [
          "RELATED",
          "RECURRENCE",
          "VERIFICATION"
        ],
        "title": "IncidentEvidenceRelationshipType",
        "type": "string"
      },
      "IncidentEvidenceResponse": {
        "additionalProperties": false,
        "properties": {
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/IncidentEvidenceItem"
            },
            "title": "Evidence",
            "type": "array"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "evidence"
        ],
        "title": "IncidentEvidenceResponse",
        "type": "object"
      },
      "IncidentStatus": {
        "enum": [
          "OPEN",
          "VERIFYING",
          "VERIFIED",
          "RECURRED"
        ],
        "title": "IncidentStatus",
        "type": "string"
      },
      "IncidentSummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "due_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due State"
          },
          "due_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Time"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "machine_id": {
            "format": "uuid",
            "title": "Machine Id",
            "type": "string"
          },
          "owner_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Ref"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "machine_id",
          "status",
          "owner_ref",
          "severity",
          "due_state",
          "due_time",
          "created_at",
          "updated_at"
        ],
        "title": "IncidentSummaryResponse",
        "type": "object"
      },
      "JsonValue": {},
...
      "MoveEvidenceRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "target_incident_id": {
            "format": "uuid",
            "title": "Target Incident Id",
            "type": "string"
          }
        },
        "required": [
          "target_incident_id",
          "reason"
        ],
        "title": "MoveEvidenceRequest",
        "type": "object"
      },
      "MoveEvidenceResponse": {
        "additionalProperties": false,
        "properties": {
          "evidence_id": {
            "format": "uuid",
            "title": "Evidence Id",
            "type": "string"
          },
          "new_link_id": {
            "format": "uuid",
            "title": "New Link Id",
            "type": "string"
          },
          "old_link_id": {
            "format": "uuid",
            "title": "Old Link Id",
            "type": "string"
          },
          "source_incident_id": {
            "format": "uuid",
            "title": "Source Incident Id",
            "type": "string"
          },
          "target_incident_id": {
            "format": "uuid",
            "title": "Target Incident Id",
            "type": "string"
          }
        },
        "required": [
          "source_incident_id",
          "target_incident_id",
          "evidence_id",
          "old_link_id",
          "new_link_id"
        ],
        "title": "MoveEvidenceResponse",
        "type": "object"
      },
...
      "SplitIncidentRequest": {
        "additionalProperties": false,
        "properties": {
          "evidence_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "evidence_ids",
          "reason"
        ],
        "title": "SplitIncidentRequest",
        "type": "object"
      },
      "SplitIncidentResponse": {
        "additionalProperties": false,
        "properties": {
          "evidence_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "new_incident_id": {
            "format": "uuid",
            "title": "New Incident Id",
            "type": "string"
          },
          "new_link_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "New Link Ids",
            "type": "array"
          },
          "old_link_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Old Link Ids",
            "type": "array"
          },
          "source_incident_id": {
            "format": "uuid",
            "title": "Source Incident Id",
            "type": "string"
          }
        },
        "required": [
          "source_incident_id",
          "new_incident_id",
          "evidence_ids",
          "old_link_ids",
          "new_link_ids"
        ],
        "title": "SplitIncidentResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MINE-TRACE Backend",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/incidents": {
      "get": {
        "operationId": "list_incidents_api_v1_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "machine_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Machine Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IncidentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "owner_ref",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Ref"
            }
          },
          {
            "in": "query",
            "name": "due_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Due State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentCollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Incidents",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "operationId": "get_incident_api_v1_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident",
        "tags": [
          "Incidents"
        ]
      }
    },
...
    "/api/v1/incidents/{incident_id}/audit": {
      "get": {
        "operationId": "get_incident_audit_api_v1_incidents__incident_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentAuditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident Audit",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/evidence": {
      "get": {
        "operationId": "get_incident_evidence_api_v1_incidents__incident_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentEvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident Evidence",
        "tags": [
          "Incidents"
        ]
      }
    },
...
    "/api/v1/incidents/{source_incident_id}/evidence/{evidence_id}/move": {
      "post": {
        "operationId": "move_incident_evidence_api_v1_incidents__source_incident_id__evidence__evidence_id__move_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Incident Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveEvidenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoveEvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Move Incident Evidence",
        "tags": [
          "Incidents"
        ]
      }
    },
    "/api/v1/incidents/{source_incident_id}/split": {
      "post": {
        "operationId": "split_incident_api_v1_incidents__source_incident_id__split_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SplitIncidentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SplitIncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Split Incident",
        "tags": [
          "Incidents"
        ]
      }
    },
...
  },
...
}